refactor: address review feedback on building hours editor
- Replace isDirty/submitted flags with JSON.stringify comparison
  for correct undo semantics (hawkrives, copilot-reviewer)
- Change reducer default case to throw instead of returning the
  action value, preventing state corruption (copilot-reviewer)
- Extract buildingReducer and BuildingAction to building-reducer.ts
  to decouple tests from UI dependencies (copilot-reviewer)
